HCBE

HCBE.edge

HCBE.build_bg

HCBE.edge_comp

HCBE.h2txt

HCBE.my_mod

HCBE.partition

HCBE.edge — Type

edge

Edge of bipartite graph with star expansion of a hypergraph.

Arguments

from : Vertex number that the edge connects to. It is a number representing a vertex of the hypergraph.

to : Vertex number that the edge connects to. It is a number representing a hyperedge of the hypergraph.

weight : Edge weight.

id : The number of edges that this structure represents.

HCBE.build_bg — Function

buildbg(h::Hypergraph, weightedf::Function=tfidf, param=Dict())

Star expand h and construct a bipartite graph. Each edge is weighted by weighted_f.

HCBE.edge_comp — Method

edge_comp(a::edge, b::edge)

Edge comparison function. Sort in descending order of weight. If the weights are the same, they are compared according to the IDs of the edges.

HCBE.h2txt — Method

h2txt(h::Hypergraph, fname::AbstractString)

Translate hypergraph to text file. This is in the following form.

N M

N is number of hypergraph vertices. M is number of hypereedges.

HCBE.my_mod — Method

my_mod(h::Hypergraph, part::Array{Set{any}})

Computes the modularity when h is divided by part.

HCBE.partition — Function

partition(uf::UnionFind{Int}, to::Int=length(uf.parent), from::Int=1, )::Vector{Set{Int}}

Compute the clustering result from the disjoint set.

Arguments

uf : The disjoint set obtained by clustering.

to : Disjoint sets up to A are used.

For example, if you want to get a vertex-only clustering result, specify the number of vertices.

from : Use a disjoint set from A.